Splinter the too-large __init__.py
