+++ grouphug/__main__.py
@@ -1,5 +1,5 @@
 
-from grouphug import app
+from grouphug.web import app
 
 if __name__ == '__main__':
     app.run(debug=True)
